refactor(cost-analysis): split codes
